+++ src/nodes/conditionals.js
@@ -128,6 +128,17 @@ const printConditional = keyword => (path, { inlineConditionals }, print) => {
 };
 
 module.exports = {
+  else: (path, opts, print) => {
+    const stmts = path.getValue().body[0];
+
+    return concat([
+      stmts.body.length === 1 && stmts.body[0].type === "command"
+        ? breakParent
+        : "",
+      "else",
+      indent(concat([softline, path.call(print, "body", 0)]))
+    ]);
+  },
   elsif: (path, opts, print) => {
     const [_predicate, _statements, addition] = path.getValue().body;
     const parts = [