+++ src/nodes/methods.js
@@ -1,4 +1,13 @@
-const { concat, group, hardline, indent } = require("../builders");
+const {
+  align,
+  concat,
+  group,
+  hardline,
+  indent,
+  join,
+  line
+} = require("../builders");
+const { literal } = require("../utils");
 
 const printMethod = offset => (path, opts, print) => {
   const [_name, params, body] = path.getValue().body.slice(offset);
@@ -38,5 +47,50 @@ const printMethod = offset => (path, opts, print) => {
 
 module.exports = {
   def: printMethod(0),
-  defs: printMethod(2)
+  defs: printMethod(2),
+  methref: (path, opts, print) => join(".:", path.map(print, "body")),
+  return: (path, opts, print) => {
+    const args = path.getValue().body[0].body[0];
+
+    if (!args) {
+      return "return";
+    }
+
+    if (args.body[0] && args.body[0].type === "paren") {
+      // Ignoring the parens node and just going straight to the content
+      return concat([
+        "return ",
+        path.call(print, "body", 0, "body", 0, "body", 0, "body", 0)
+      ]);
+    }
+
+    return concat(["return ", join(", ", path.call(print, "body", 0))]);
+  },
+  return0: literal("return"),
+  super: (path, opts, print) => {
+    const args = path.getValue().body[0];
+
+    if (args.type === "arg_paren") {
+      // In case there are explicitly no arguments but they are using parens,
+      // we assume they are attempting to override the initializer and pass no
+      // arguments up.
+      if (args.body[0] === null) {
+        return "super()";
+      }
+
+      return concat(["super", path.call(print, "body", 0)]);
+    }
+
+    return concat(["super ", join(", ", path.call(print, "body", 0))]);
+  },
+  undef: (path, opts, print) =>
+    group(
+      concat([
+        "undef ",
+        align(
+          "undef ".length,
+          join(concat([",", line]), path.map(print, "body", 0))
+        )
+      ])
+    )
 };